% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_coding_genes.R
\docType{data}
\name{cached_coding_genes}
\alias{cached_coding_genes}
\title{cached_coding_genes.RData}
\format{A \code{chr} vector of gene symbols for all protein coding genes}
\source{
ftp://ftp.ebi.ac.uk/pub/databases/genenames/new/tsv/locus_groups/protein-coding_gene.txt
}
\usage{
cached_coding_genes
}
\value{
Chr vector of gene symbols
}
\description{
Gene symbols for all human protein coding genes, cached for use in RITAN vignettes. The set of all human protein coding genes is the default "background" list. This data was generated by running the load_all_protein_coding_symbols() function from RITAN and saving the resulting vector of gene symbols. By default, functions in RITAN will download the most recent list. We encourage users to use the most appropriate background for their study. See our vignettes for additional discussion.
}
\examples{
print(str(cached_coding_genes))
}
\author{
Michael T Zimmermann
}
\keyword{datasets}

