% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_list.R
\docType{data}
\name{geneset_list}
\alias{geneset_list}
\title{geneset_list.RData}
\format{A \code{list} with one element for each annotation resource. Each element contains a varying number of gene symbols.}
\source{
Multiple sources used. See our table in the network vignette.
}
\usage{
geneset_list
}
\value{
A list of resources. Each resource contains a list of genesets.
}
\description{
R list object containing annotation resources and the gene sets contained in each of them.
}
\examples{
print(names(geneset_list))
}
\author{
Michael T Zimmermann
}
\keyword{datasets}

