% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_list.R
\docType{data}
\name{network_list}
\alias{network_list}
\title{network_list.RData}
\format{A \code{list} of resources, one entry for each network resource. Each list element is a data.frame with 3 columns:
\describe{
 \item{\code{p1}}{starting point}
 \item{\code{edge_type}}{connection type}
 \item{\code{p2}}{ending point}
 }}
\source{
Multiple sources used. See our table in the network vignette.
}
\usage{
network_list
}
\value{
A \code{list} of netowrk resources. Each entry contains a \code{list} of edges as a \code{data.frame}.
}
\description{
R list object containing 15 annotation resources and the gene sets contained in each of them.
}
\examples{
print(names(network_list))
}
\author{
Michael T Zimmermann
}
\keyword{datasets}
