% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_coding_genes.R
\docType{data}
\name{cached_coding_genes}
\alias{cached_coding_genes}
\title{cached_coding_genes.RData}
\format{A \code{chr} vector of gene symbols for all protein coding genes}
\source{
ftp://ftp.ebi.ac.uk/pub/databases/genenames/new/tsv/locus_groups/protein-coding_gene.txt
}
\usage{
cached_coding_genes
}
\value{
Chr vector of gene symbols
}
\description{
gene symbols for protein coding genes - cached for use in vignettes -
}
\examples{
print(str(cached_coding_genes))
}
\author{
Michael T Zimmermann
}
\keyword{datasets}
